requirements.txt
# Basic dependencies for stock search AI project
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0

# OpenAI SDK with agents support
openai>=1.35.0
openai-agents

# Alpaca Trading API
alpaca-py
